+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/PatternRendering.java
@@ -1,9 +1,7 @@
 package com.refinedmods.refinedstorage.common.autocrafting;
 
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
-import com.refinedmods.refinedstorage.common.api.autocrafting.CraftingPattern;
 import com.refinedmods.refinedstorage.common.api.autocrafting.PatternProviderItem;
-import com.refinedmods.refinedstorage.common.api.autocrafting.ProcessingPattern;
 import com.refinedmods.refinedstorage.common.support.resource.ItemResource;
 import com.refinedmods.refinedstorage.common.util.PlatformUtil;
 
@@ -26,17 +24,19 @@ public final class PatternRendering {
         if (level == null) {
             return Optional.empty();
         }
-        return RefinedStorageApi.INSTANCE.getPattern(stack, level).map(pattern -> {
-            if (pattern instanceof CraftingPattern craftingPattern
-                && craftingPattern.output().resource() instanceof ItemResource itemResource) {
-                return itemResource.toItemStack();
-            }
-            if (pattern instanceof ProcessingPattern processingPattern
-                && processingPattern.outputs().size() == 1
-                && processingPattern.outputs().getFirst().resource() instanceof ItemResource itemResource) {
-                return itemResource.toItemStack();
-            }
-            return null;
+        return RefinedStorageApi.INSTANCE.getPattern(stack, level).map(pattern -> switch (pattern) {
+            case CraftingPattern craftingPattern
+                when craftingPattern.output().resource() instanceof ItemResource itemResource ->
+                itemResource.toItemStack();
+            case ProcessingPattern processingPattern
+                when processingPattern.outputs().size() == 1
+                && processingPattern.outputs().getFirst().resource() instanceof ItemResource itemResource ->
+                itemResource.toItemStack();
+            case StonecutterPattern stonecutterPattern
+                when stonecutterPattern.output() instanceof ItemResource itemResource -> itemResource.toItemStack();
+            case SmithingTablePattern smithingTablePattern
+                when smithingTablePattern.output() instanceof ItemResource itemResource -> itemResource.toItemStack();
+            default -> null;
         });
     }
 }