+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/StonecutterPattern.java
@@ -1,0 +1,7 @@
+package com.refinedmods.refinedstorage.common.autocrafting;
+
+import com.refinedmods.refinedstorage.common.api.autocrafting.Pattern;
+import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
+
+record StonecutterPattern(PlatformResourceKey input, PlatformResourceKey output) implements Pattern {
+}